Initial Project Design
This is initial commit includes a semi-functional demo to show the basic
idea of the project.

Includes:
MediaPlayerModelBuilder.java to set the primary stage of the
application.
MediaPlayerModel.java to hold the necessary values for the application
Model.
MediaPlayerController.java to handle the functionality of the
application.
MediaPlayerView.fxml to outline the basic look of the application's GUI.
MediaPlayerStyle.css to contain all GUI style changes for the
application.