Updated Project Design and Added Functionality
Renamed MediaPlayerModel to MediaPlayerPlayList and switch it to only
handle the user's current playlist.
Renamed MediaPlayerModelBuilder to MediaPlayerModel to be more accurate.
Editted MediaPlayerView to allow for a dynamic display field.
Added functionality to MediaPlayerController that allows the user to
continuously play media from a playlist.
Added functionality to MediaPlayerController that allows the user to
skip forward and backward through a playlist.
Added two temporary media files and a temporary .bin file to allow
testing of the application.